adding docker-compose

+++ main.sh
@@ -50,7 +50,7 @@ echo -e "  ${YELLOW}3.Install bt Panel${RES}                            ${YELLOW
 echo ""
 echo -e "  ${YELLOW}5.Modify SSH Port (for defaut port 22)${RES}        ${YELLOW}6.Unixbech一键跑分${RES}"
 echo ""
-echo -e "  ${YELLOW}7.Install docker${RES}                              ${YELLOW}8.Upgrade sqlite3${RES}" 
+echo -e "  ${YELLOW}7.Install docker and docker-compose${RES}                              ${YELLOW}8.Upgrade sqlite3${RES}" 
 echo ""
 echo -e "  ${YELLOW}9.Set localtime to China zone${RES}                 ${YELLOW}10.VPS info${RES}"
 echo ""
@@ -87,10 +87,27 @@ elif [ "$main_no" = "5" ]; then
 elif [ "$main_no" = "6" ]; then
 	wget --no-check-certificate https://github.com/teddysun/across/raw/master/unixbench.sh && chmod +x unixbench.sh && ./unixbench.sh
 elif [ "$main_no" = "7" ]; then
-	curl -fsSL https://get.docker.com -o get-docker.sh && sh get-docker.sh
-	echo ""
-	echo "Please run 'service docker start' to start docker service."
-	echo "Or run 'systemctl enable docker' to make docker autostart when the system reboots."
+	read -p "do you want to install docker service? [y (default) or n] " bbb
+	if [ "$bbb" = "n"]; then
+		break;
+	else
+		curl -fsSL https://get.docker.com -o get-docker.sh && sh get-docker.sh
+		systemctl enable docker
+		systemctl start docker
+		echo ""
+		echo "docker server is working."
+		echo ""
+	fi
+	read -p "do you want to install docker-compose? [y (default) or n] " aaa
+	if [ "$aaa" = "n"]; then
+		exit 0;
+	else
+		curl -L https://github.com/docker/compose/releases/download/1.17.0/docker-compose-`uname -s`-`uname -m` -o /usr/local/bin/docker-compose
+		chmod +x /usr/local/bin/docker-compose
+		echo ""
+		echo "docker-compose has been installed."
+		echo ""
+	fi
 	echo ""
 elif [ "$main_no" = "8" ]; then
 	wget https://github.com/hityne/centos/raw/main/upgrade-sqlite3.sh && chmod a+x upgrade-sqlite3.sh && bash upgrade-sqlite3.sh
